DDF-136: implemented Pre and Post Federated Query plugins; added min/max values for RRD data sources to suppress spike data; changed Long values to Double for CSV, XLS, JSON, and XML generated reports

+++ metrics/metrics-reporting/src/test/java/ddf/metrics/reporting/internal/rrd4j/RrdMetricsRetrieverTest.java
@@ -198,7 +198,7 @@ public class RrdMetricsRetrieverTest extends XMLTestCase
         for (int i=0; i < samples.size(); i++)
         {
             JSONObject sample = (JSONObject) samples.get(i);
-            LOGGER.debug("timestamp = " + (String) sample.get("timestamp") + ",   value = " + (Long) sample.get("value"));
+            LOGGER.debug("timestamp = " + (String) sample.get("timestamp") + ",   value = " + sample.get("value"));
             assertThat(sample.get("timestamp"), not(nullValue()));
             assertThat(sample.get("value"), not(nullValue()));
         }
